implemented new data file format

+++ src/SimpleMockServer.Domain.DataModel/DataImpls/Float/FloatData.cs
@@ -1,8 +1,8 @@
 ﻿namespace SimpleMockServer.Domain.DataModel.DataImpls.Float;
 
-public class FloatData : Data<float>
+public class FloatData : Data<decimal>
 {
-    public FloatData(DataName name, IReadOnlyDictionary<DataName, DataRange<float>> ranges) :
+    public FloatData(DataName name, IReadOnlyDictionary<DataName, DataRange<decimal>> ranges) :
         base(name, ranges.ToDictionary(x => x.Key, x => x.Value))
     {
     }
