+++ src/SimpleMockServer.Domain.DataModel/DataImpls/Float/FloatDataRange.cs
@@ -1,10 +1,10 @@
 ﻿namespace SimpleMockServer.Domain.DataModel.DataImpls.Float;
 
-public abstract class FloatDataRange : DataRange<float>
+public abstract class FloatDataRange : DataRange<decimal>
 {
     protected FloatDataRange(DataName name) : base(name)
     {
     }
 
-    public override bool TryParse(string str, out float value) => float.TryParse(str, out value);
+    public override bool TryParse(string str, out decimal value) => decimal.TryParse(str, out value);
 }
